requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.1.1
Deprecated==1.2.12
Django==3.2.5
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
jwcrypto==0.9.1
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
